requirements_nano.txt
absl-py==0.9.0
appdirs==1.4.4
astor==0.8.1
certifi==2020.4.5.1
chardet==3.0.4
cvlib==0.2.5
cycler==0.10.0
Cython==0.29.17
decorator==4.4.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
imageio==2.8.0
importlib-metadata==1.6.0
imutils==0.5.3
joblib==0.15.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.16.1
opt-einsum==3.2.1
Pillow==7.1.2
progressbar==2.5
protobuf==3.11.3
pycuda==2019.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytools==2020.2
PyYAML==5.3.1
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.1.0
six==1.14.0
sklearn==0.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu @ https://developer.download.nvidia.com/compute/redist/jp/v42/tensorflow-gpu/tensorflow_gpu-1.15.0+nv19.11-cp36-cp36m-linux_aarch64.whl
termcolor==1.1.0
threadpoolctl==2.0.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
